+++ packages/layout/src/components/PageContainer/index.tsx
@@ -1,13 +1,17 @@
 import type { GenerateStyle } from '@ant-design/pro-provider';
 import { ProConfigProvider, ProProvider } from '@ant-design/pro-provider';
-import type {
+import {
+  Affix,
   AffixProps,
+  Breadcrumb,
   BreadcrumbProps,
+  ConfigProvider,
   SpinProps,
   TabPaneProps,
+  Tabs,
   TabsProps,
+  version,
 } from 'antd';
-import { Affix, Breadcrumb, ConfigProvider, Tabs } from 'antd';
 import classNames from 'classnames';
 import type { ReactNode } from 'react';
 import React, { useContext, useEffect, useMemo } from 'react';
@@ -24,6 +28,8 @@ import { WaterMark } from '../WaterMark';
 import type { PageContainerToken, pageContainerToken } from './style';
 import { useStyle } from './style';
 import { useStylish } from './style/stylish';
+
+import { compareVersions } from '@ant-design/pro-utils';
 
 import 'antd/lib/breadcrumb/style';
 
@@ -150,7 +156,6 @@ const renderFooter: React.FC<
           }
         }}
         tabBarExtraContent={tabBarExtraContent}
-        // @ts-ignore
         items={tabList?.map((item, index) => ({
           label: item.tab,
           ...item,
@@ -158,11 +163,18 @@ const renderFooter: React.FC<
         }))}
         {...tabProps}
       >
-        {tabList?.map((item, index) => {
-          return (
-            <Tabs.TabPane key={item.key || index} tab={item.tab} {...item} />
-          );
-        })}
+        {/* 如果版本低于 4.23.0，不支持 items */}
+        {compareVersions(version, '4.23.0') < 0
+          ? tabList?.map((item, index) => {
+              return (
+                <Tabs.TabPane
+                  key={item.key || index}
+                  tab={item.tab}
+                  {...item}
+                />
+              );
+            })
+          : null}
       </Tabs>
     );
   }