+++ packages/table/src/components/ListToolBar/index.tsx
@@ -1,7 +1,13 @@
 import { proTheme, useIntl } from '@ant-design/pro-provider';
-import { LabelIconTip } from '@ant-design/pro-utils';
-import type { TabPaneProps } from 'antd';
-import { ConfigProvider, Input, Tabs, Tooltip } from 'antd';
+import { compareVersions, LabelIconTip } from '@ant-design/pro-utils';
+import {
+  ConfigProvider,
+  Input,
+  TabPaneProps,
+  Tabs,
+  Tooltip,
+  version,
+} from 'antd';
 import type { LabelTooltipType } from 'antd/lib/form/FormItemLabel';
 import type { SearchProps } from 'antd/lib/input';
 import classNames from 'classnames';
@@ -117,7 +123,6 @@ const ListToolBarTabBar: React.FC<{
           }}
           defaultActiveKey={tabs.defaultActiveKey}
           activeKey={tabs.activeKey}
-          //@ts-ignore
           items={tabs.items.map((item, index) => ({
             label: item.tab,
             ...item,
@@ -127,9 +132,9 @@ const ListToolBarTabBar: React.FC<{
           tabBarExtraContent={filtersNode}
         >
           {tabs.items?.map((item, index) => {
-            return (
+            return compareVersions(version, '4.23.0') < 0 ? (
               <Tabs.TabPane {...item} key={item.key || index} tab={item.tab} />
-            );
+            ) : null;
           })}
         </Tabs>
       ) : (