+++ packages/components/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.23](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.6.22...@ant-design/pro-components@2.6.23) (2023-09-20)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
+## [2.6.22](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.6.21...@ant-design/pro-components@2.6.22) (2023-09-19)
+
+**Note:** Version bump only for package @ant-design/pro-components
+
 ## [2.6.21](https://github.com/ant-design/pro-components/compare/@ant-design/pro-components@2.6.20...@ant-design/pro-components@2.6.21) (2023-09-18)
 
 **Note:** Version bump only for package @ant-design/pro-components