+++ packages/field/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.12.8](https://github.com/ant-design/pro-components/compare/@ant-design/pro-field@2.12.7...@ant-design/pro-field@2.12.8) (2023-09-19)
+
+**Note:** Version bump only for package @ant-design/pro-field
+
 ## [2.12.7](https://github.com/ant-design/pro-components/compare/@ant-design/pro-field@2.12.6...@ant-design/pro-field@2.12.7) (2023-09-18)
 
 ### Bug Fixes