+++ packages/form/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.18.10](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.18.9...@ant-design/pro-form@2.18.10) (2023-09-19)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
 ## [2.18.9](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@2.18.8...@ant-design/pro-form@2.18.9) (2023-09-18)
 
 ### Bug Fixes